Refactorisation de la MainActivy en une ListActivity. Mise à jour des méthodes et importations.
